Add a test to the rev-list `command`
In this test we use `git rev-list` and a specific algorithm in
`ocaml-git` used by the push command to calculate what we need to send
to the server.

* we follow how `git` calculate what it needs to send but we don't
  generate the PACK file
* from what we know, we calculate what `git rev-list` calculate and
  generate the PACK file

Then we compare the radix tree produced by the `ocaml-git`'s packer
and compare it with what `git rev-list` said.

To know how git generate a PACK file (and calculate, btw, list of
objects), you can see `send-pack.c` and specifically `pack_objects`
function which uses `rev-list`'s syntax (revision) to calculate a set of
objects to prepare them to be packed.