Add rev-list test on git-unix package
This test in __only__ available on the git-unix package because we
compare results with what git said - so we need a usual git context with
a .git directory. Obviously, we did not have any diff between the
memory/mirage back-end and the unix back-end about this algorithm
(rev-list). So, if it's work on git-unix, it should be work on
memory/mirage!